pyproject.toml
[project]
name = "e-commerce-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt==4.0.1",
    "chromadb>=0.6.3",
    "cloudinary>=1.44.0",
    "email-validator==2.1.0.post1",
    "fastapi==0.105.0",
    "livekit-agents[cartesia,deepgram,openai,silero,turn-detector]~=1.0",
    "livekit-plugins-noise-cancellation~=0.2",
    "motor==3.3.1",
    "open-clip-torch>=2.32.0",
    "passlib==1.7.4",
    "pillow>=11.2.1",
    "pydantic-settings>=2.9.1",
    "pymongo==4.6.0",
    "python-dotenv>=1.1.0",
    "python-jose==3.3.0",
    "python-multipart==0.0.6",
    "razorpay>=1.4.2",
    "setuptools>=80.4.0",
    "uvicorn==0.24.0",
]
